+++ branches/avidemux_2.7/avidemux_core/ADM_coreAudio/src/ADM_audioStreamEac3.cpp
@@ -16,7 +16,7 @@ ADM_audioStreamEAC3::ADM_audioStreamEAC3(WAVHeader *header,ADM_audioAccess *acce
     if(access->canGetDuration()==false)
     {
         // We can compute the duration from the length
-        float size=access->getLength();
+        float size=(float)access->getLength();
               size/=header->byterate; // Result is in second
               size*=1000;
               size*=1000; // s->us
